added postfix configuration capability

+++ jammy/install.sh
@@ -1,14 +1,21 @@
 #!/bin/bash
 # Ubuntu Pro token from: https://ubuntu.com/pro/dashboard (not needed for Ubuntu Pro instances on Azure, AWS, or Google Cloud)
 TOKEN=''
+FQDN=voip.yourdomain.com
+HOSTNAME=voip
 DOMAIN=yourdomain.com
-SENDGRID_API_KEY=your-api-key
-GMAIL_USERNAME=username@gmail.com
-GMAIL_PASSWORD=yourpassword
-# toggle to decide if sendgrid or gmail configs will be applied
+SMTP_HOST=smtp.gmail.com
+SMTP_PORT=587
+SMTP_USERNAME=username@gmail.com
+SMTP_PASSWORD=yourpassword
+PRETTY_HOSTNAME="My PBX Server"
+sudo hostnamectl set-hostname "$FQDN" --static
+sudo hostnamectl set-hostname "$FQDN" --transient
+sudo hostnamectl set-hostname "$HOSTNAME"
+sudo hostnamectl set-hostname "$PRETTY_HOSTNAME" --pretty
 bash -c 'cat <<EOF > /etc/systemd/system/freepbx.service
 [Unit]
-Description=FreePBX VoIP Server
+Description=$PRETTY_HOSTNAME
 After=mariadb.service
 [Service]
 Type=oneshot
@@ -29,56 +36,28 @@ DEBIAN_FRONTEND=noninteractive apt-get install -y libapache2-mod-php7.4
 debconf-set-selections <<< "postfix postfix/mailname string $DOMAIN"
 debconf-set-selections <<< "postfix postfix/main_mailer_type string 'Internet Site'"
 DEBIAN_FRONTEND=noninteractive apt-get install -y postfix
-# Configure server to send email via Sendgrid
-echo $DOMAIN > /etc/mailname
-# Check the length of the Sendgrid API Key, they have to be 69 characters
-if [ ${#var} -eq 69 ]; then
-# configure main.cf for sendgrid
-sed -i.bak -E 's/^([ \t]*append_dot_mydomain[ \t]*=[ \t]*).*/\1'"yes"'/' /etc/postfix/main.cf
-sed -i.bak -E 's/^([ \t]*myhostname[ \t]*=[ \t]*).*/\1'"$DOMAIN"'/' /etc/postfix/main.cf
-sed -i.bak -E 's/^([ \t]*mydestination[ \t]*=[ \t]*).*/\1'"localhost"'/' /etc/postfix/main.cf
-sed -i.bak -E 's/^([ \t]*relayhost[ \t]*=[ \t]*).*/\1'"[smtp.sendgrid.net]:2525"'/' /etc/postfix/main.cf
-bash -c 'cat <<EOF >> /etc/postfix/main.cf
-smtp_tls_security_level = encrypt
-smtp_sasl_auth_enable = yes
-smtp_sasl_password_maps = hash:/etc/postfix/sasl_passwd
-header_size_limit = 4096000
-smtp_sasl_security_options = noanonymous
-EOF'
-# write the sendgrid API key to the sasl_passwd file
-echo [smtp.sendgrid.net]:2525 apikey:$SENDGRID_API_KEY >> /etc/postfix/sasl_passwd
-else
-# Configure server to send email via Gmail
-cat << EOF > /etc/postfix/main.cf
-# Basic configuration
-myhostname = $DOMAIN
-myorigin = $DOMAIN
-mydestination = localhost
-relayhost = [smtp.gmail.com]:587
-inet_interfaces = loopback-only
-
-# Security configuration
-smtp_sasl_auth_enable = yes
-smtp_sasl_security_options = noanonymous
-smtp_sasl_password_maps = hash:/etc/postfix/sasl_passwd
-smtp_tls_security_level = encrypt
-smtp_tls_CAfile = /etc/ssl/certs/ca-certificates.crt
-EOF
-# Create the sasl_passwd file to store the Gmail login information
-echo "[smtp.gmail.com]:587 username@gmail.com:password" > /etc/postfix/sasl_passwd
+if [[ "${SMTP_HOST,,}" == "smtp.gmail.com" ]]; then
+  wget https://www.thawte.com/roots/thawte_Primary_Root_CA.pem -O /etc/postfix/thawte_Primary_Root_CA.pem
+  chmod 400 /etc/postfix/thawte_Primary_Root_CA.pem
+  postconf -e smtp_tls_CAfile=/etc/postfix/thawte_Primary_Root_CA.pem
+  postconf -e smtp_use_tls=yes
 fi
-
-# Use the postmap utility to generate a .db file, and set appropriate privileges
-postmap /etc/postfix/sasl_passwd
-chown root:root /etc/postfix/sasl_passwd /etc/postfix/sasl_passwd.db
-chmod 600 /etc/postfix/sasl_passwd.db
-
-# remove file containing credentials
-rm /etc/postfix/sasl_passwd
-
-# Reload your configuration to load the modified parameters:
-/etc/init.d/postfix restart
-
+if [[ -n "${SMTP_HOST}" ]]; then
+  postconf -e myhostname="${FQDN}"
+  postconf -e mydomain="${DOMAIN}"
+  postconf -e default_transport=smtp
+  postconf -e relay_transport=smtp
+  postconf -e relayhost="[$SMTP_HOST]:${SMTP_PORT}"
+  postconf -e smtp_tls_security_level=encrypt
+  postconf -e smtp_sasl_auth_enable=yes
+  postconf -e smtp_sasl_password_maps=hash:/etc/postfix/sasl_passwd
+  postconf -e header_size_limit=4096000
+  postconf -e smtp_sasl_security_options=noanonymous
+  postmap /etc/postfix/sasl_passwd
+  rm /etc/postfix/sasl_passwd
+  chmod 600 /etc/postfix/sasl_passwd.db
+  /etc/init.d/postfix restart
+fi
 DEBIAN_FRONTEND=noninteractive apt-get install -y curl dirmngr ffmpeg git lame libicu-dev mpg123 sqlite3 sox
 # `/usr/src/asterisk-18*/contrib/scripts/install_prereq install` identifies these packages
 DEBIAN_FRONTEND=noninteractive apt-get install -y bison doxygen flex graphviz libcfg-dev libcorosync-common-dev libcpg-dev libjack-jackd2-dev libldap2-dev libosptk-dev pkgconf subversion xmlstarlet
